refactor(morphology): optimize participle analysis and flatten case enums
- Optimized `src/morphology/participle.rs` by introducing `std::sync::LazyLock` to pre-compute and sort `ALL_PATTERNS`. This eliminates repeated vector allocation and sorting on every call to `analyze_participle`, significantly improving performance.
- Flattened the module hierarchy by moving grammatical enums (`Case`, `Number`, `Gender`, `Person`, `Tense`, `Mood`, `Voice`) from `src/morphology/case.rs` to `src/morphology/mod.rs`.
- Deleted `src/morphology/case.rs` as it is no longer needed.
- Updated `.jules/razor.md` with reduction metrics.

This aligns with the Razor philosophy of simplifying code and removing bloat.
